Remove `NEW_` prefixes and split plugin templates
