+++ avalanche/training/supervised/strategy_wrappers.py
@@ -34,7 +34,7 @@ from avalanche.training.plugins import (
     MASPlugin,
 )
 from avalanche.training.templates.base import BaseTemplate
-from avalanche.training.templates.supervised import SupervisedTemplate
+from avalanche.training.templates.common_templates import SupervisedTemplate
 from avalanche.models.generator import MlpVAE, VAE_loss
 from avalanche.logging import InteractiveLogger
