+++ avalanche/training/templates/__init__.py
@@ -11,6 +11,8 @@ Templates are the backbone that supports the plugin systems.
 """
 from .base import BaseTemplate
 from .base_sgd import BaseSGDTemplate
-from .base_online_sgd import BaseOnlineSGDTemplate
-from .online_supervised import OnlineSupervisedTemplate
-from .supervised import SupervisedTemplate
+from .common_templates import (
+    SupervisedTemplate,
+    SupervisedMetaLearningTemplate,
+    OnlineSupervisedTemplate
+)
